requirements.txt
python-dotenv==1.0.0
streamlit==1.25.0
sqlite3==0.0.1
google-generativeai==0.1.0
SQLAlchemy==2.0.0
